feature: add confirmation message on deleting

+++ src/app/components/tasks/tasks.component.ts
@@ -18,6 +18,8 @@ export class TasksComponent implements OnInit {
   }
 
   deleteTask(task: Task) {
-    this.taskService.deleteTask(task).subscribe(() => this.tasks = this.tasks.filter(t => t.id !== task.id));
+    if (confirm('Are you sure you wanna delete this task?')) {
+      this.taskService.deleteTask(task).subscribe(() => this.tasks = this.tasks.filter(t => t.id !== task.id));
+    }
   }
 }
